condor_annex version 0.0.4
The "Runtime" : "nodejs" parameter is no longer supported for creating
and/or updating AWS Lambda functions. If condor_annex is executed with
this deprecated runtime, the ActivateAlarm, CleanUp, and Plumbing Lambda
functions are not created, causing the annex's CloudFormation Stack to
rollback. The new "Runtime" : "nodejs4.3" is now used while creating or
updating these functions.