condor_annex version 0.0.5
Removed the s3:deleteObject action on the S3PoolPassword and
S3ConfigFile in the lambda_cleanup_policy found in
CloudFormation/template-2. This prevents deletion of the pool password
and user configuration files stored in a static S3 bucket when the
--password-url and --config-url command-line options are used. Note,
however, this may lead the deletion of the S3DeleteBucket to fail when
the --password-file and --config-file options are used instead. i.e.,
the S3 bucket dynamically created with the --password-file and
--config-file options will no longer be non-empty when the annex's
CloudFormation stack attempts to clean itself up and shutdown.